+++ ms-pedidos/src/main/java/cl/duoc/cloud/pedidos/controller/dto/NuevoPedidoDTO.java
@@ -1,4 +1,4 @@
 package cl.duoc.cloud.pedidos.controller.dto;
 
-public record NuevoPedidoDTO(String cliente, Long productoId, int cantidad) {
+public record NuevoPedidoDTO(Long productoId, int cantidad) {
 }